Add the components to the Main screen
Add the Accompanist Insets dependency.
Wrap the Main content in the ProvideWindowInsets function to setup
Insets in the composables.
Add a Scaffold to the Main composable to implement the basic material
design visual layout structure.
Create the MainNavGraph file.
Define Main destinations in the MainNavGraph file.
Add the detailed components in the design to the Main screen.